LayoutTests/imported/w3c/web-platform-tests/resource-timing/nextHopProtocol-is-tao-protected.https-expected.txt


FAIL Add TAO-less iframe from remote origin. Make sure nextHopProtocol is the empty string assert_equals: nextHopProtocol should be the empty string expected "" but got "http/1.1"
PASS Add TAO'd iframe from remote origin. Make sure nextHopProtocol is not the empty string
